Adjust includes, types, and string usage
Normalize Prey include paths, fix string/logging calls, and address pointer/size type issues for portability and 64-bit correctness. Many ../prey headers were changed to ../Prey to match directory casing. Logging/Error/Warning calls now pass std::string::c_str() where needed to avoid format issues. Pointer arithmetic and stored pointers were converted to uintptr_t (and related casts) and alignment checks were tightened in idHeap and Model_lwo to fix 64-bit/ alignment bugs. Removed obsolete register usage and corrected sizeof usage in MD5_Final, adjusted huft_build locals, and updated jpeg extern prototypes to return void. Minor fixes include correcting a member name in prey_interpolate and ensuring final newline; installer binaries were also updated.

+++ src/renderer/Model_lwo.cpp
@@ -2179,7 +2179,7 @@ int lwGetPolygons5( idFile *fp, int cksize, lwPolygonList *plist, int ptoffset )
          bp += 2;
       }
       j -= 1;
-      pp->surf = ( lwSurface * ) j;
+      pp->surf = ( lwSurface * )(uintptr_t)j;
 
       pp++;
       pv += nv;
@@ -2696,7 +2696,7 @@ int lwResolvePolySurfaces( lwPolygonList *polygon, lwTagList *tlist,
    }
 
    for ( i = 0; i < polygon->count; i++ ) {
-      index = ( int ) polygon->pol[ i ].surf;
+      index = ( int )(uintptr_t)polygon->pol[ i ].surf;
       if ( index < 0 || index > tlist->count ) return 0;
       if ( !s[ index ] ) {
          s[ index ] = lwDefaultSurface();
@@ -2876,7 +2876,7 @@ int lwGetPolygonTags( idFile *fp, int cksize, lwTagList *tlist, lwPolygonList *p
 		if ( rlen < 0 || rlen > cksize ) return 0;
 
 		switch ( type ) {
-			case ID_SURF:  plist->pol[ i ].surf = ( lwSurface * ) j;  break;
+			case ID_SURF:  plist->pol[ i ].surf = ( lwSurface * )(uintptr_t)j;  break;
 			case ID_PART:  plist->pol[ i ].part = j;  break;
 			case ID_SMGP:  plist->pol[ i ].smoothgrp = j;  break;
 		}
